+++ packages/framework/src/Models/Pages/MarkdownPost.php
@@ -23,7 +23,6 @@ class MarkdownPost extends AbstractMarkdownPage
     public function __construct(string $identifier = '', ?FrontMatter $matter = null, ?Markdown $markdown = null)
     {
         parent::__construct($identifier, $matter, $markdown);
-        $this->constructMetadata();
     }
 
     protected function constructPageSchemas(): void
@@ -47,55 +46,5 @@ class MarkdownPost extends AbstractMarkdownPage
     public static function getLatestPosts(): Collection
     {
         return static::all()->sortByDesc('matter.date');
-    }
-
-    // HasArticleMetadata (Generates article metadata for a MarkdownPost)
-
-    /** @deprecated pending move to service */
-    public array $postMetadata = [];
-    /** @deprecated pending move to service */
-    public array $properties = [];
-
-    protected function constructMetadata(): void
-    {
-        $this->parseFrontMatterMetadata();
-
-        $this->makeOpenGraphPropertiesForArticle();
-    }
-
-    /** @deprecated pending move to service */
-    public function getPostMetadata(): array
-    {
-        return $this->postMetadata;
-    }
-
-    /** @deprecated pending move to service */
-    public function getMetaProperties(): array
-    {
-        return $this->properties;
-    }
-
-    /**
-     * Generate metadata from the front matter that can be used in standard <meta> tags.
-     * This helper is page type agnostic and works with any kind of model having front matter.
-     *
-     * @deprecated pending move to service
-     */
-    protected function parseFrontMatterMetadata(): void
-    {
-    }
-
-    /**
-     * Generate opengraph metadata from front matter for an og:article such as a blog post.
-     *
-     * @deprecated pending move to service
-     */
-    protected function makeOpenGraphPropertiesForArticle(): void
-    {
-    }
-
-    /** @deprecated pending move to service */
-    protected function setImageMetadata(): void
-    {
     }
 }